@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "im-data-manager-workflow-engine"
-version = "0.1.0"
+version = "1.0.0"
 description = "The Data Manager workflow engine, schema and decoder."
 authors = [
     "Alan Christie <achristie@informaticsmatters.com>",
@@ -12,7 +12,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.12"
+python = "^3.13"
 im-protobuf = "^8.2.0"
 im-data-manager-job-decoder = "^2.1.0"
 jsonschema = "^4.21.1"
