@@ -4,7 +4,7 @@ This is typically used by the Data Manager's Workflow Engine.
 """
 
 import os
-from typing import Any, Dict, Optional
+from typing import Any
 
 import jsonschema
 import yaml
@@ -19,11 +19,11 @@ _WORKFLOW_SCHEMA_FILE: str = os.path.join(
 # This must work as the file is installed along with this module.
 assert os.path.isfile(_WORKFLOW_SCHEMA_FILE)
 with open(_WORKFLOW_SCHEMA_FILE, "r", encoding="utf8") as schema_file:
-    _WORKFLOW_SCHEMA: Dict[str, Any] = yaml.load(schema_file, Loader=yaml.FullLoader)
+    _WORKFLOW_SCHEMA: dict[str, Any] = yaml.load(schema_file, Loader=yaml.FullLoader)
 assert _WORKFLOW_SCHEMA
 
 
-def validate_schema(workflow: Dict[str, Any]) -> Optional[str]:
+def validate_schema(workflow: dict[str, Any]) -> str | None:
     """Checks the Workflow Definition against the built-in schema.
     If there's an error the error text is returned, otherwise None.
     """
